weather-app/requirements.txt
requests
kafka-python-ng
JsonSerializer